+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## v1.0.0 / 202x MMM DD
+## v1.1.0 / 2021 Oct 21
+
+> This release adds the `:suppress-bugsnag-response?` option to `notify` to make behavior more consistent with other logging interfaces
+> Additionally include a `notify-v2!` which surpresses this output by default
+
+## v1.0.0 / 2021 Oct 19
 
 > This release officially forks com.splashfinancial/clj-bugsnag from the original project